+++ src/cli/commands/run-once/pipeline-legacy.ts
@@ -1,0 +1,953 @@
+import { join } from "node:path";
+import { localPiAgentDir } from "../init/pi-agent-settings.ts";
+
+import { createRunOnceHostProvider } from "../../../host/factory.ts";
+import { planLabelChange } from "../triage/labels.ts";
+import { materializeIssueArtifactSources } from "./artifact-source-materialization.ts";
+import { assertApprovedArtifactsResolvable } from "./approval-artifact-preflight.ts";
+import { runArtifactSourceStage } from "./artifact-source-stage.ts";
+import type { ResolvedIssueArtifactSources } from "./artifact-sources.ts";
+import { ensureAutomationLabel } from "./automation-labels.ts";
+import {
+  assertCleanWorktree,
+  assertIssueBaseContainedInPrBase,
+  ensureIssueWorktree,
+  inspectIssueWorkspace,
+  type IssueWorktreeResult,
+} from "./git.ts";
+import {
+  PlanningArtifactSafetyError,
+  resolvePlanningArtifacts,
+  type PlanningArtifactPolicy,
+} from "./planning-artifacts.ts";
+import { advancePlanningStages } from "./stage-advancement.ts";
+import {
+  ApprovalRequiredError,
+  cleanupLabelsForImplementation,
+  resolveWorkflowState,
+} from "./workflow-state.ts";
+import {
+  isResumableRunState,
+  readRunState,
+  replaceRunStateAfterReset,
+  writeRunState,
+} from "./run-state.ts";
+import { selectIssueWithDiagnostics } from "./selection.ts";
+import {
+  createStepAccounting,
+  emitSimpleStep,
+  progress,
+  withLogPath,
+} from "./pipeline-progress.ts";
+import {
+  AgentIssueSafetyError,
+  hasBlockedRunRecoveryState,
+  effectiveCheckpoints,
+  lifecycleLabels,
+  recoveryClaimLabels,
+  resetReceiptCheckpoints,
+  nextLabels,
+  workflowTransition,
+} from "./pipeline-lifecycle.ts";
+import {
+  blockerComment,
+  blockerCommentKey,
+  startedComment,
+} from "./pipeline-comments.ts";
+import {
+  cleanStatusIgnoredPaths,
+  configuredWorktreeStrategy,
+  expectedIssueWorkspace,
+  planningArtifactPolicyForWorkspace,
+} from "./pipeline-workspace.ts";
+import {
+  emitSelectionDiagnostics,
+  loadSelectionIssues,
+  selectResumableIssue,
+} from "./pipeline-selection.ts";
+import { blockIssue, unexpectedFailure } from "./pipeline-failures.ts";
+import { withIssueRunLease } from "./recovery-lease.ts";
+import { formatRunRecoveryDecision } from "./recovery.ts";
+import {
+  adoptLegacyRecoveryLease,
+  recoverBlockedWorkspace,
+} from "./pipeline-recovery.ts";
+import { runPipelineImplementationStage } from "./pipeline-implementation.ts";
+import { runPipelineFinishStage } from "./pipeline-finish.ts";
+import { resolvePipelineRunCost } from "./pipeline-run-cost.ts";
+import {
+  runPiSessionPath,
+  type AgentIssueProgressEvent,
+  type ProgressReporter,
+} from "./progress.ts";
+import type {
+  AgentIssueConfig,
+  AgentIssuePipelineResult,
+  AgentIssueRunState,
+  CommandRunner,
+  IssueSummary,
+} from "./types.ts";
+
+type PiOutputStream = (chunk: string) => void;
+
+export type RunOneIssueOptions = {
+  now?: Date | undefined;
+  progress?: ProgressReporter | undefined;
+  logPath?: string | undefined;
+  streamPiOutput?: PiOutputStream | undefined;
+  verbosePiOutput?: boolean | undefined;
+  heartbeatMs?: number | undefined;
+};
+type LeasedRunOneIssueOptions = RunOneIssueOptions & {
+  lease?: import("./types.ts").IssueRunLease;
+  /** Internal selection pin; never exposed to ordinary callers. */
+  leasedIssueNumber?: number;
+  reset?: { seed: import("./types.ts").RunResetSeed };
+};
+
+function hasFinishedPlanningWorkspaceState(
+  state: AgentIssueRunState | undefined,
+): boolean {
+  return (
+    state?.status === "finished" &&
+    state.implementationStatus === undefined &&
+    !!(state.specPath || state.planPath) &&
+    !!(state.branch || state.worktreePath)
+  );
+}
+
+export async function runLegacyOneIssue(
+  runner: CommandRunner,
+  config: AgentIssueConfig,
+  options: RunOneIssueOptions = {},
+): Promise<AgentIssuePipelineResult> {
+  return runLegacyOneIssueInternal(runner, config, options);
+}
+
+/** Reset uses this narrow leased entry point after it has archived state. */
+export async function runLegacyOneIssueAfterReset(
+  runner: CommandRunner,
+  config: AgentIssueConfig,
+  options: RunOneIssueOptions,
+  reset: {
+    lease: import("./types.ts").IssueRunLease;
+    seed: import("./types.ts").RunResetSeed;
+  },
+): Promise<AgentIssuePipelineResult> {
+  return runLegacyOneIssueInternal(runner, config, {
+    ...options,
+    lease: reset.lease,
+    reset: { seed: reset.seed },
+  });
+}
+
+async function runLegacyOneIssueInternal(
+  runner: CommandRunner,
+  config: AgentIssueConfig,
+  options: LeasedRunOneIssueOptions = {},
+): Promise<AgentIssuePipelineResult> {
+  const host = createRunOnceHostProvider({
+    runner,
+    repoRoot: config.repoRoot,
+    host: config.host,
+  });
+  // Once an automatic selection has acquired its lease, re-read only that
+  // issue. Never perform another priority selection while holding a lease for
+  // a different issue.
+  const loadedIssues =
+    options.leasedIssueNumber === undefined
+      ? await loadSelectionIssues(host, config, options)
+      : [await host.viewIssue(options.leasedIssueNumber)];
+  // A blocked saved attempt is never an implicit retry target.  It needs an
+  // explicit issue acknowledgement and the normal recovery gate below.
+  const issues =
+    config.issueNumber === undefined
+      ? (
+          await Promise.all(
+            loadedIssues.map(async (candidate) => ({
+              candidate,
+              state: await readRunState(config.runStateDir, candidate.number),
+            })),
+          )
+        )
+          .filter(({ state }) => !hasBlockedRunRecoveryState(state))
+          .map(({ candidate }) => candidate)
+      : loadedIssues;
+  let selected: { issue: IssueSummary; resumed: boolean } | undefined;
+  try {
+    selected = options.reset
+      ? (() => {
+          const resetIssue = issues.find(
+            (candidate) => candidate.number === config.issueNumber,
+          );
+          if (!resetIssue)
+            throw new Error(
+              `Reset issue #${config.issueNumber ?? "?"} disappeared before pipeline entry`,
+            );
+          return { issue: resetIssue, resumed: false };
+        })()
+      : await selectResumableIssue(issues, config);
+  } catch (error) {
+    if (error instanceof ApprovalRequiredError) {
+      return withLogPath(
+        {
+          status: "approval-required",
+          issue: error.issue,
+          approvalKind: error.approvalKind,
+          missingLabel: error.missingLabel,
+        },
+        options,
+      );
+    }
+    throw error;
+  }
+  const issue = selected?.issue;
+
+  if (!issue) {
+    if (config.issueNumber === undefined) {
+      const diagnostics = selectIssueWithDiagnostics(issues, {
+        readyLabel: lifecycleLabels(config).ready,
+        triagePolicy: config.triagePolicy,
+        approvalPolicy: config.approvalPolicy,
+      });
+      await emitSelectionDiagnostics(diagnostics.rejections, options);
+      await progress(
+        options,
+        "info",
+        "select",
+        `no eligible issue found after considering ${diagnostics.consideredCount} open issues; see run log for skip details`,
+      );
+    } else {
+      await progress(options, "info", "select", "no eligible issue found");
+    }
+    return withLogPath({ status: "no-issue" }, options);
+  }
+
+  // A reset has already archived and validated the prior attempt.  Its seed is
+  // authoritative; do not let pre-reset checkpoints re-enter resume policy.
+  if (options.lease && options.lease.record.issueNumber !== issue.number)
+    throw new AgentIssueSafetyError(
+      `Run lease for issue #${options.lease.record.issueNumber} cannot process issue #${issue.number}`,
+    );
+  if (
+    options.leasedIssueNumber !== undefined &&
+    options.leasedIssueNumber !== issue.number
+  )
+    throw new AgentIssueSafetyError(
+      `Leased selection changed from issue #${options.leasedIssueNumber} to issue #${issue.number}`,
+    );
+
+  let existingState: AgentIssueRunState | undefined = options.reset
+    ? {
+        issueNumber: issue.number,
+        title: issue.title,
+        status: "claimed",
+        ...(options.reset.seed.specPath
+          ? { specPath: options.reset.seed.specPath }
+          : {}),
+        ...(options.reset.seed.specCommit
+          ? { specCommit: options.reset.seed.specCommit }
+          : {}),
+        ...(options.reset.seed.planPath
+          ? { planPath: options.reset.seed.planPath }
+          : {}),
+        ...(options.reset.seed.planCommit
+          ? { planCommit: options.reset.seed.planCommit }
+          : {}),
+        checkpoints: resetReceiptCheckpoints(options.reset.seed),
+        leaseProtocolVersion: 1,
+        createdAt: (options.now ?? new Date()).toISOString(),
+        updatedAt: (options.now ?? new Date()).toISOString(),
+      }
+    : await readRunState(config.runStateDir, issue.number);
+  // Every real attempt is serialized from selection through its final effect.
+  // A caller-supplied lease (reset) is borrowed and is therefore not released.
+  if (!config.dryRun && !options.lease) {
+    return withIssueRunLease(
+      { runStateDir: config.runStateDir, issueNumber: issue.number },
+      (lease) =>
+        runLegacyOneIssueInternal(runner, config, {
+          ...options,
+          lease,
+          leasedIssueNumber: issue.number,
+        }),
+    );
+  }
+  if (!config.dryRun && options.lease)
+    existingState = await adoptLegacyRecoveryLease({
+      config,
+      issueNumber: issue.number,
+      state: existingState,
+      lease: options.lease,
+    });
+  const piAgentDir = localPiAgentDir(config.repoRoot);
+  const resumed = selected?.resumed ?? false;
+  const ordinaryResumableState =
+    resumed && !!existingState && isResumableRunState(existingState);
+
+  await progress(
+    options,
+    "info",
+    "select",
+    `${resumed ? "resuming" : "selected"} #${issue.number} ${issue.title}`,
+    { issueNumber: issue.number },
+  );
+  await progress(
+    options,
+    "info",
+    "git",
+    "checking issue branch base containment",
+    { issueNumber: issue.number },
+  );
+  await assertIssueBaseContainedInPrBase(
+    runner,
+    config.repoRoot,
+    config.baseRef,
+    config.remote,
+    config.baseBranch,
+  );
+  const timestamp = (options.now ?? new Date()).toISOString();
+  const piSessionPath = runPiSessionPath(
+    config.runStateDir,
+    timestamp,
+    issue.number,
+  );
+  const runOptions = { ...options, piSessionPath };
+  if (config.dryRun) {
+    const { ready } = lifecycleLabels(config);
+    const state = resolveWorkflowState(issue.labels, {
+      readyLabel: ready,
+      policy: config.approvalPolicy,
+    });
+    return withLogPath(
+      {
+        status: "dry-run",
+        issue,
+        transition: workflowTransition(state, config),
+      },
+      runOptions,
+    );
+  }
+
+  const ignoredPaths = cleanStatusIgnoredPaths(
+    config,
+    runOptions.logPath === undefined ? {} : { logPath: runOptions.logPath },
+  );
+  // Typed recovery owns every blocked-state classification, including safely
+  // recreatable missing worktrees and branches.  Do not pre-gate it here.
+  const blockedRecoveryResumable = hasBlockedRunRecoveryState(existingState);
+  const planningWorkspaceResumable =
+    hasFinishedPlanningWorkspaceState(existingState);
+  const resumableState =
+    ordinaryResumableState ||
+    blockedRecoveryResumable ||
+    planningWorkspaceResumable;
+  const resetStaleCheckpoints = !!existingState && !resumableState;
+  const checkpoints = {
+    ...(options.reset
+      ? resetReceiptCheckpoints(options.reset.seed)
+      : (effectiveCheckpoints(existingState?.checkpoints, resumableState) ??
+        {})),
+  };
+
+  if (
+    resetStaleCheckpoints &&
+    (existingState?.branch || existingState?.worktreePath)
+  ) {
+    throw new AgentIssueSafetyError(
+      `Non-resumable run state for issue #${issue.number} has stale branch/worktree; clean up before starting a fresh run`,
+    );
+  }
+
+  const tokenUsageState = { total: 0 };
+  let issueForRun = issue;
+  let resolvedArtifacts: ResolvedIssueArtifactSources = {};
+  const stepAccounting = createStepAccounting({
+    progress: runOptions.progress,
+    issueNumber: issue.number,
+    runStartedAtMs: (runOptions.now ?? new Date()).getTime(),
+  });
+  const stepStart = stepAccounting.start;
+  const stepComplete = stepAccounting.complete;
+  const runStep = stepAccounting.run;
+  const observePi =
+    (
+      stage:
+        | "pi-artifact-extraction"
+        | "pi-plan"
+        | "pi-development-environment"
+        | "pi-implementation",
+    ) =>
+    async (
+      observation: AgentIssueProgressEvent["observation"],
+    ): Promise<void> => {
+      await stepAccounting.observe(stage, observation);
+    };
+  await runOptions.progress?.event({
+    time: timestamp,
+    level: "info",
+    stage: "run",
+    message: `issue #${issue.number} · ${issue.title}`,
+    issueNumber: issue.number,
+    step: { type: "run-start", issueNumber: issue.number, title: issue.title },
+  });
+  await emitSimpleStep(runOptions, issue.number, "select issue");
+  const artifactSources = await runArtifactSourceStage({
+    host,
+    config,
+    issue,
+    now: runOptions.now ?? new Date(),
+    progress: (level, stage, message, extras) =>
+      progress(runOptions, level, stage, message, extras),
+    runStep,
+  });
+  issueForRun = artifactSources.issue;
+  resolvedArtifacts = artifactSources.resolvedArtifacts;
+  let specPath: string | undefined;
+  let specCommit: string | undefined;
+  let planPath: string | undefined;
+  let planCommit: string | undefined;
+  let branch: string | undefined;
+  let worktreePath: string | undefined;
+  let ensuredWorktree: IssueWorktreeResult | undefined;
+  let artifactPolicy: PlanningArtifactPolicy | undefined;
+  const worktreeStrategy = configuredWorktreeStrategy(config);
+  const expectedWorkspace = expectedIssueWorkspace(
+    issue.number,
+    issue.title,
+    worktreeStrategy,
+  );
+  const blockedRecovery = await recoverBlockedWorkspace({
+    runner,
+    config,
+    issueNumber: issue.number,
+    existingState,
+    expectedWorkspace,
+    ignoredPaths,
+    resolvedArtifacts,
+    lease: options.lease,
+  });
+  if (
+    blockedRecovery?.decision.action === "resume" &&
+    blockedRecovery.decision.assessment.worktree.ignoredEntries.length
+  ) {
+    const message = formatRunRecoveryDecision(blockedRecovery.decision);
+    await progress(runOptions, "info", "recovery", message, {
+      issueNumber: issue.number,
+      consoleMessage: message,
+      data: {
+        action: "resume",
+        mutationApplied: false,
+        ignoredEntries:
+          blockedRecovery.decision.assessment.worktree.ignoredEntries,
+      },
+    });
+  }
+  const assertExpectedWorkspaceIdentity = (): void => {
+    if (
+      resumableState &&
+      existingState?.branch &&
+      existingState.branch !== expectedWorkspace.branch
+    ) {
+      throw new AgentIssueSafetyError(
+        `Saved branch ${existingState.branch} does not match expected branch ${expectedWorkspace.branch}`,
+      );
+    }
+    if (
+      resumableState &&
+      existingState?.worktreePath &&
+      existingState.worktreePath !== expectedWorkspace.worktreePath
+    ) {
+      throw new AgentIssueSafetyError(
+        `Saved worktree ${existingState.worktreePath} does not match expected worktree path ${expectedWorkspace.worktreePath}`,
+      );
+    }
+  };
+  const hasApprovalLabel = [
+    config.approvalPolicy.specApproval.approvedLabel,
+    config.approvalPolicy.planApproval.approvedLabel,
+  ].some((label) => issueForRun.labels.includes(label));
+  const artifactWorkspace = hasApprovalLabel
+    ? await (async () => {
+        assertExpectedWorkspaceIdentity();
+        return await inspectIssueWorkspace(runner, config.repoRoot, {
+          branch: expectedWorkspace.branch,
+          worktreePath:
+            existingState?.worktreePath ?? expectedWorkspace.worktreePath,
+        });
+      })()
+    : undefined;
+  const ensureIssueWorkspace = async (): Promise<IssueWorktreeResult> => {
+    if (ensuredWorktree) return ensuredWorktree;
+    assertExpectedWorkspaceIdentity();
+
+    const worktree = await ensureIssueWorktree(
+      runner,
+      config.repoRoot,
+      issue.number,
+      issue.title,
+      worktreeStrategy,
+      undefined,
+      ignoredPaths,
+    );
+    await emitSimpleStep(
+      runOptions,
+      issue.number,
+      worktree.created ? "create worktree" : "reuse worktree",
+    );
+    if (
+      resumableState &&
+      existingState?.branch &&
+      existingState.branch !== worktree.branch
+    ) {
+      throw new AgentIssueSafetyError(
+        `Saved branch ${existingState.branch} does not match ensured worktree branch ${worktree.branch}`,
+      );
+    }
+    if (
+      resumableState &&
+      existingState?.worktreePath &&
+      existingState.worktreePath !== worktree.worktreePath
+    ) {
+      throw new AgentIssueSafetyError(
+        `Saved worktree ${existingState.worktreePath} does not match ensured worktree path ${worktree.worktreePath}`,
+      );
+    }
+    branch = resumableState
+      ? (existingState?.branch ?? worktree.branch)
+      : worktree.branch;
+    worktreePath = resumableState
+      ? (existingState?.worktreePath ?? worktree.worktreePath)
+      : worktree.worktreePath;
+    await progress(
+      runOptions,
+      "info",
+      "worktree",
+      `${worktree.created ? "creating" : "reusing"} worktree ${worktreePath}`,
+      { issueNumber: issue.number },
+    );
+    ensuredWorktree = worktree;
+    return worktree;
+  };
+
+  const { ready, inProgress, done, needsInfo } = lifecycleLabels(config);
+  const approvedArtifactPreflight = await assertApprovedArtifactsResolvable({
+    config,
+    issue: issueForRun,
+    existingState,
+    resolvedArtifacts,
+    now: runOptions.now ?? new Date(),
+    artifactWorkspace,
+    runner,
+  });
+  await progress(runOptions, "info", "git", "checking repository status", {
+    issueNumber: issue.number,
+  });
+  await assertCleanWorktree(runner, config.repoRoot, ignoredPaths);
+  if (artifactWorkspace?.kind === "branch") {
+    const workspace = await ensureIssueWorkspace();
+    artifactPolicy = planningArtifactPolicyForWorkspace({
+      config,
+      existingState,
+      resolvedArtifacts,
+      worktreePath: workspace.worktreePath,
+      allowGeneratedSpec: false,
+      allowGeneratedPlan: false,
+    });
+  } else if (approvedArtifactPreflight) {
+    artifactPolicy = approvedArtifactPreflight.policy;
+  }
+
+  const recoveringBlocked =
+    hasBlockedRunRecoveryState(existingState) || !!options.reset;
+  let labels = recoveryClaimLabels(issue.labels, {
+    ready,
+    inProgress,
+    needsInfo,
+    recoveringBlocked,
+  });
+  if (
+    recoveringBlocked ||
+    !checkpoints.claimed ||
+    (planningWorkspaceResumable && !issue.labels.includes(inProgress))
+  ) {
+    await runStep("claim issue", async () => {
+      await progress(
+        runOptions,
+        "info",
+        "labels",
+        `ensuring ${inProgress} label exists`,
+        { issueNumber: issue.number },
+      );
+      await ensureAutomationLabel(host, config, inProgress);
+      await host.applyLabels(
+        planLabelChange(issue.number, issue.labels, labels),
+      );
+      await progress(
+        runOptions,
+        "info",
+        "claim",
+        `claimed #${issue.number}: ${ready} -> ${inProgress}`,
+        { issueNumber: issue.number },
+      );
+      if (options.reset) {
+        await replaceRunStateAfterReset(
+          config.runStateDir,
+          {
+            issueNumber: issue.number,
+            title: issue.title,
+            seed: options.reset.seed,
+          },
+          timestamp,
+        );
+      } else {
+        if (recoveringBlocked && existingState?.lastError) {
+          const blocked = {
+            status: "blocked" as const,
+            reason: existingState.lastError,
+            questions: existingState.blockerQuestions ?? [],
+            commits: existingState.commits ?? [],
+            validation: existingState.validation ?? [],
+          };
+          const body = blockerComment(blocked);
+          if (issue.comments?.some((comment) => comment.body === body))
+            await writeRunState(
+              config.runStateDir,
+              {
+                issueNumber: issue.number,
+                title: issue.title,
+                status: existingState.status,
+                blockerCommentKeys: [blockerCommentKey(blocked)],
+              },
+              timestamp,
+            );
+        }
+        await writeRunState(
+          config.runStateDir,
+          {
+            issueNumber: issue.number,
+            title: issue.title,
+            status: "claimed",
+            checkpoints: { claimed: true },
+            resetCheckpoints: resetStaleCheckpoints,
+            clearLastError: recoveringBlocked,
+            clearBlockerQuestions: recoveringBlocked,
+            leaseProtocolVersion: 1,
+          },
+          timestamp,
+        );
+      }
+      checkpoints.claimed = true;
+    });
+  }
+
+  try {
+    if (!checkpoints.startedCommentPosted) {
+      await host.commentIssue(issueForRun.number, startedComment(issueForRun));
+      await writeRunState(
+        config.runStateDir,
+        {
+          issueNumber: issueForRun.number,
+          title: issueForRun.title,
+          status: "claimed",
+          checkpoints: { startedCommentPosted: true },
+        },
+        timestamp,
+      );
+      checkpoints.startedCommentPosted = true;
+    }
+
+    if (
+      !artifactPolicy &&
+      resumableState &&
+      existingState &&
+      (existingState.branch || existingState.worktreePath) &&
+      !!(existingState.specPath || existingState.planPath)
+    ) {
+      const resumeWorktree = await ensureIssueWorkspace();
+      const savedWorktreePath =
+        existingState.worktreePath ?? resumeWorktree.worktreePath;
+      artifactPolicy = planningArtifactPolicyForWorkspace({
+        config,
+        existingState,
+        resolvedArtifacts,
+        worktreePath: savedWorktreePath,
+        allowGeneratedSpec: false,
+        allowGeneratedPlan: false,
+      });
+      await progress(
+        runOptions,
+        "info",
+        "resume",
+        `reusing saved worktree for artifact lookup: ${savedWorktreePath}`,
+        { issueNumber: issue.number },
+      );
+    }
+
+    if (artifactPolicy?.kind === "implementation-resume") {
+      const resumeArtifactPolicy = artifactPolicy;
+      const sourcesToMaterialize = {
+        ...(!resumeArtifactPolicy.saved.specPath && resolvedArtifacts.spec
+          ? { spec: resolvedArtifacts.spec }
+          : {}),
+        ...(!artifactPolicy.saved.planPath && resolvedArtifacts.plan
+          ? { plan: resolvedArtifacts.plan }
+          : {}),
+      };
+      if (sourcesToMaterialize.spec || sourcesToMaterialize.plan) {
+        const materializedArtifacts = await runStep(
+          "materialize issue artifact sources",
+          async () =>
+            materializeIssueArtifactSources({
+              repoRoot: resumeArtifactPolicy.primary.repoRoot,
+              runner,
+              issueNumber: issueForRun.number,
+              sources: sourcesToMaterialize,
+            }),
+        );
+        resolvedArtifacts = {
+          ...resolvedArtifacts,
+          ...materializedArtifacts,
+        };
+        artifactPolicy = {
+          ...artifactPolicy,
+          explicit: resolvedArtifacts,
+        };
+      }
+      if (!approvedArtifactPreflight) {
+        await resolvePlanningArtifacts({
+          policy: artifactPolicy,
+          issue: issueForRun,
+          now: runOptions.now ?? new Date(),
+        });
+      }
+    } else {
+      const artifactWorktree =
+        resolvedArtifacts.spec || resolvedArtifacts.plan
+          ? await ensureIssueWorkspace()
+          : undefined;
+      const artifactRepoRoot = artifactWorktree
+        ? join(config.repoRoot, artifactWorktree.worktreePath)
+        : config.repoRoot;
+      const materializedArtifacts = await runStep(
+        "materialize issue artifact sources",
+        async () =>
+          materializeIssueArtifactSources({
+            repoRoot: artifactRepoRoot,
+            runner,
+            issueNumber: issueForRun.number,
+            sources: resolvedArtifacts,
+          }),
+      );
+      resolvedArtifacts = materializedArtifacts;
+    }
+
+    if (!artifactPolicy && (resolvedArtifacts.spec || resolvedArtifacts.plan)) {
+      await writeRunState(
+        config.runStateDir,
+        {
+          issueNumber: issueForRun.number,
+          title: issueForRun.title,
+          status: "planning",
+          specPath: resolvedArtifacts.spec?.path,
+          specCommit: resolvedArtifacts.spec?.commit,
+          planPath: resolvedArtifacts.plan?.path,
+          planCommit: resolvedArtifacts.plan?.commit,
+        },
+        timestamp,
+      );
+    }
+
+    const planningStages = await advancePlanningStages({
+      runner,
+      host,
+      config,
+      issue: issueForRun,
+      labels,
+      ready,
+      inProgress,
+      needsInfo,
+      approvalGatesSatisfied:
+        ordinaryResumableState && existingState?.status === "implementing",
+      existingState,
+      resolvedArtifacts,
+      artifactPolicy,
+      ensurePlanningArtifactWorkspace: async () => {
+        const workspace = await ensureIssueWorkspace();
+        return {
+          repoRoot: join(config.repoRoot, workspace.worktreePath),
+          branch: workspace.branch,
+          worktreePath: workspace.worktreePath,
+        };
+      },
+      checkpoints,
+      timestamp,
+      now: runOptions.now ?? new Date(),
+      runOptions,
+      piAgentDir,
+      tokenUsageState,
+      progress: (level, stage, message, extras) =>
+        progress(runOptions, level, stage, message, extras),
+      runStep,
+      observePi,
+      emitSimpleStep: (issueNumber, label) =>
+        emitSimpleStep(runOptions, issueNumber, label),
+      blockIssue: (result, details) =>
+        blockIssue(
+          host,
+          config,
+          issueForRun,
+          labels,
+          result,
+          details,
+          timestamp,
+          runOptions,
+        ),
+    });
+    if (planningStages.kind === "finished") {
+      return withLogPath(planningStages.result, runOptions);
+    }
+
+    labels = planningStages.labels;
+    specPath = planningStages.specPath;
+    specCommit = planningStages.specCommit;
+    planPath = planningStages.planPath;
+    planCommit = planningStages.planCommit;
+
+    const implementationLabels = nextLabels(
+      cleanupLabelsForImplementation(labels, {
+        readyLabel: ready,
+        policy: config.approvalPolicy,
+      }),
+      [],
+      [inProgress],
+    );
+    if (implementationLabels.join("\0") !== labels.join("\0")) {
+      await host.applyLabels(
+        planLabelChange(issue.number, labels, implementationLabels),
+      );
+      labels = implementationLabels;
+    }
+
+    const worktree = await ensureIssueWorkspace();
+    const implementationWorktreePath = worktreePath ?? worktree.worktreePath;
+    const implementationBranch = branch ?? worktree.branch;
+    if (!implementationBranch || !planPath) {
+      throw new AgentIssueSafetyError(
+        `Implementation requires a branch and plan for issue #${issue.number}`,
+      );
+    }
+    const implementationStage = await runPipelineImplementationStage({
+      runner,
+      host,
+      config,
+      issue,
+      labels,
+      readyLabel: ready,
+      inProgressLabel: inProgress,
+      specPath,
+      specCommit,
+      planPath,
+      planCommit,
+      branch: implementationBranch,
+      worktreePath: implementationWorktreePath,
+      worktree,
+      worktreeStrategy,
+      existingState,
+      resumableState,
+      implementationCompleted: checkpoints.implementationCompleted,
+      checkpoints,
+      timestamp,
+      runOptions,
+      piAgentDir,
+      tokenUsageState,
+      progressReporter: runOptions.progress,
+      runStep,
+      stepStart,
+      stepComplete,
+      observePi,
+      emitSimpleStep: (issueNumber, label) =>
+        emitSimpleStep(runOptions, issueNumber, label),
+      blockIssue: (result, details) =>
+        blockIssue(
+          host,
+          config,
+          issue,
+          labels,
+          result,
+          details,
+          timestamp,
+          runOptions,
+        ),
+    });
+
+    if (implementationStage.kind === "blocked") {
+      return withLogPath(implementationStage.result, runOptions);
+    }
+    if (implementationStage.kind === "unexpected") {
+      throw implementationStage.error;
+    }
+    labels = implementationStage.labels;
+    checkpoints.worktreeReady = true;
+
+    const runCostReport = await resolvePipelineRunCost({
+      implementationKind: implementationStage.kind,
+      implementationStatus: implementationStage.result.status,
+      piSessionPath: runOptions.piSessionPath,
+      persistedReport: existingState?.runCostReport,
+      warn: (message, error) =>
+        progress(runOptions, "warning", "run-cost", message, {
+          issueNumber: issue.number,
+          data: error instanceof Error ? error.message : String(error ?? ""),
+          consoleMessage: `Warning: ${message}`,
+        }),
+    });
+
+    const finishStage = await runPipelineFinishStage({
+      runner,
+      host,
+      config,
+      issue,
+      labels,
+      readyLabel: ready,
+      inProgressLabel: inProgress,
+      doneLabel: done,
+      needsInfoLabel: needsInfo,
+      checkpoints,
+      implemented: implementationStage.result,
+      runCostReport,
+      specPath,
+      specCommit,
+      planPath,
+      planCommit,
+      branch: implementationBranch,
+      worktreePath: implementationWorktreePath,
+      timestamp,
+      runOptions,
+      runStep,
+    });
+
+    if (finishStage.kind === "unexpected") {
+      throw finishStage.error;
+    }
+
+    return finishStage.result;
+  } catch (error) {
+    if (error instanceof AgentIssueSafetyError) {
+      throw error;
+    }
+    if (error instanceof PlanningArtifactSafetyError) {
+      throw new AgentIssueSafetyError(error.message);
+    }
+    return unexpectedFailure(
+      host,
+      config,
+      issue,
+      checkpoints,
+      { specPath, specCommit, planPath, planCommit, branch, worktreePath },
+      timestamp,
+      error,
+      runOptions,
+    );
+  }
+}
